requirements.txt
flask
psycopg2-binary
web3
tensorflow